Changed behaviour: The filename to vardok should be in the first line of tdok from now on. This first line gets removed then. It is also possible to create several files now: for each line in vardok. This will be done when the first column in vardok is called FILENAME.

+++ templateReplacer.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-if ARGV.length != 2
+if ARGV.length != 1
 	File.open("README.md", "r") do |infile|
 	    while (line = infile.gets)
 	        puts "#{line}"
@@ -12,18 +12,24 @@ end
 
 tdok = ARGV[0]
 if not File.exist?(tdok)
-	puts "Die Datei #{tdok} existiert nicht!"
+	puts "The file #{tdok} does not exist in #{Dir.pwd}..."
 	exit
 end
 
-vardok = ARGV[1]
+# read path to vardok
+readfile = File.new(tdok, "r")
+vardok = readfile.gets.chomp
+readfile.close
+
 if not File.exist?(vardok)
-	puts "Die Datei #{vardok} existiert nicht!"
+	puts "The file #{vardok} does not exist in #{Dir.pwd}... Check the first line of tdok (#{tdok})."
 	exit
 end
 
 varnames = Array.new
 vars = Array.new
+
+mode = "replacements in one file below each other"
 
 # Werte holen
 begin
@@ -34,6 +40,11 @@ begin
 	while (line = readfile.gets)
 		if first_line
 			varnames = line.strip.split(";;;")
+
+			if varnames[0] == "FILENAME"
+				mode = "replacements in new file always"
+			end
+
 			first_line = false
 			next
 		end
@@ -56,21 +67,44 @@ tmpdok = ""
 # Werte einsetzen
 begin
 	readfile = File.new(tdok, "r")
+	first = true
 	while (line = readfile.gets)
-		dok += line
+		if first
+			# there is just the filename of vardok in this first line
+			first = false
+		else
+			dok += line
+		end
 	end
 	readfile.close
 
-	File.open("#{tdok}.replaced", 'w') do |writefile|
-		vars.each_with_index do |v, index|
-			tmpdok = String.new(dok)
+	if mode == "replacements in one file below each other"
+		File.open("#{tdok}.replaced", 'w') do |writefile|
+			vars.each_with_index do |v, index|
+				tmpdok = String.new(dok)
 
-			varnames.each_with_index do |name, index2|
-				tmpdok.gsub!("*** #{name} ***", v[index2] ? v[index2] : "")
+				varnames.each_with_index do |name, index2|
+					tmpdok.gsub!("*** #{name} ***", v[index2] ? v[index2] : "")
+				end
+
+				writefile.write(tmpdok + "\n\n")
 			end
-
-			writefile.write(tmpdok + "\n\n")
 		end
+	elsif mode == "replacements in new file always"
+		vars.each_with_index do |v, index|
+			File.open(v[0], 'w') do |writefile|
+				tmpdok = String.new(dok)
+
+				# skip first element (it's the FILENAME) - that's why we have to add 1 as well
+				varnames[1..varnames.size].each_with_index do |name, index2|
+					tmpdok.gsub!("*** #{name} ***", v[index2 + 1] ? v[index2 + 1] : "")
+				end
+
+				writefile.write(tmpdok)
+			end
+		end
+	else
+		puts "unknown mode ... exit"
 	end
 rescue => err
 	puts "Exception: #{err}"
